Updated to build script

+++ init-build.sh
@@ -6,6 +6,7 @@ PROJECT_LIBS_DIR=${PWD}
 # Initialize variables to store the values of the command-line options
 MICROKIT_APP=""
 PLATFORM=""
+BUILD_TYPE=""
 
 # Function to display usage information
 usage() {
@@ -22,6 +23,9 @@ while [ "$#" -gt 0 ]; do
         -DPLATFORM=*)
             PLATFORM="${1#*=}"
             ;;
+        -DBUILD_TYPE=*)
+            BUILD_TYPE="${1#*=}"
+            ;;
         *)
             usage
             ;;
@@ -35,16 +39,16 @@ if [ -z "$MICROKIT_APP" ] || [ -z "$PLATFORM" ]; then
     usage
 fi
 
-echo "HOST DIR: ${HOST_PATH}"
-
 echo "MICROKIT_APP: $MICROKIT_APP"
 echo "PLATFORM: $PLATFORM"
+echo "BUILD_TYPE: $BUILD_TYPE"
 
 # Go back to root dir
 cd ..
 
-# Project Path
-PROJECT_DIR="${PWD}/examples/maaxboard/$MICROKIT_APP"
+# Path's
+ROOT_DIR="${PWD}"
+PROJECT_DIR="$ROOT_DIR/examples/maaxboard/$MICROKIT_APP"
 
 # Build picolibc
 echo "Do you want to build picolibc? (yes/no)"
@@ -54,22 +58,31 @@ read user_input1
 user_input1=$(echo "$user_input1" | tr '[:upper:]' '[:lower:]')
 
 if [ "$user_input1" = "yes" ]; then
-    rm -rf picolibc/picolibc-microkit/
-    mkdir picolibc/picolibc-microkit/
-    cd picolibc/picolibc-microkit/
+    PICOLIBC_MICROKIT_DIR="picolibc/picolibc-microkit"
+    rm -rf $PICOLIBC_MICROKIT_DIR
+    mkdir $PICOLIBC_MICROKIT_DIR
+    cd $PICOLIBC_MICROKIT_DIR
     sudo rm -rf ../picolibc_build
     mkdir ../picolibc_build
-    ../scripts/do-aarch64-configure-nocrt -Dprefix=${PWD}/../../picolibc_build
+    ../scripts/do-aarch64-configure-nocrt -Dprefix=$ROOT_DIR/picolibc_build
     sudo ninja 
     sudo ninja install
 fi
 
 # Build application 
 cd $PROJECT_DIR
-rm -rf build
-mkdir build
-rm -rf example-build
-mkdir example-build
+if [ ! -d "build" ]; then
+    mkdir build
+fi
+if [ ! -d "example-build" ]; then
+    mkdir example-build
+fi
+if [ "$BUILD_TYPE" == "rebuild" ]; then
+    rm -rf build
+    mkdir build
+    rm -rf example-build
+    mkdir example-build
+fi
 cd build
 cmake -DMICROKIT_APP=$MICROKIT_APP -DPLATFORM=$PLATFORM $PROJECT_LIBS_DIR
 make
